+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- `@reykjavik/webtools/alertsStore`:
+  - feat: Bump the default `defaultDuration` to `'MEDIUM'` (more reasonable)
 
 ## 0.3.12 – 0.3.14